+++ src/compute-engine/boxed-expression/order.ts
@@ -6,7 +6,6 @@ import {
   SemiBoxedExpression,
 } from '../public';
 import { lex, maxDegree, totalDegree } from '../symbolic/polynomials';
-import { getVars } from './utils';
 
 export type Order = 'lex' | 'dexlex' | 'grevlex' | 'elim';
 
@@ -192,7 +191,7 @@ export function polynomialOrder(expr: BoxedExpression): SemiBoxedExpression {
   // Empirically, the Total Degree Reverse Lexicographic Order (grevlex)
   // is often the fastest to calculate Gröbner basis. We use it as the
   // default ordering for polynomials.
-  return degreeReverseLexicographicOrder(expr, getVars(expr));
+  return degreeReverseLexicographicOrder(expr, expr.freeVars);
 }
 
 export function lexicographicOrder(
@@ -200,7 +199,7 @@ export function lexicographicOrder(
   vars?: string[]
 ): SemiBoxedExpression {
   // @todo
-  vars = vars ?? getVars(expr);
+  vars = vars ?? expr.freeVars;
   return expr;
 }
 
@@ -209,7 +208,7 @@ export function degreeLexicographicOrder(
   vars?: string[]
 ): SemiBoxedExpression {
   // @todo
-  vars = vars ?? getVars(expr);
+  vars = vars ?? expr.freeVars;
   return expr;
 }
 
@@ -218,7 +217,7 @@ export function degreeReverseLexicographicOrder(
   vars?: string[]
 ): SemiBoxedExpression {
   // @todo
-  vars = vars ?? getVars(expr);
+  vars = vars ?? expr.freeVars;
   return expr;
 }
 
@@ -227,7 +226,7 @@ export function eliminationOrder(
   vars?: string[]
 ): SemiBoxedExpression {
   // @todo
-  vars = vars ?? getVars(expr);
+  vars = vars ?? expr.freeVars;
   return expr;
 }