+++ src/compute-engine/boxed-expression/utils.ts
@@ -48,42 +48,41 @@ export function getImaginaryCoef(expr: BoxedExpression): number | null {
 /**
  * Return the free symbols in the expression, recursively.
  * A variable, or free symbol, is a symbol that is not bound to a value.
+ * Note: do not use `isFree`: it has a side effect of creating a definition
+ * if one does not exist, and we want to avoid that. For example, `assume()`
+ * relies on `expr.freeVars` *not* creating a definition.
  */
-export function getVars(expr: BoxedExpression): string[] {
+export function getFreeVars(expr: BoxedExpression, set: Set<string>): void {
   if (expr.symbol) {
-    const def = expr.symbolDefinition;
-    return def?.constant ? [] : [expr.symbol];
+    const def = expr.engine.lookupSymbol(expr.symbol);
+    if (def?.value === undefined) set.add(expr.symbol);
+    return;
   }
 
-  if (!expr.ops && !expr.keys) return [];
+  if (!expr.ops && !expr.keys) return;
 
-  const result: string[] = [];
-
-  if (expr.ops) for (const op of expr.ops) result.push(...getVars(op));
+  if (expr.ops) for (const op of expr.ops) getFreeVars(op, set);
 
   if (expr.keys)
-    for (const key of expr.keys) result.push(...getVars(expr.getKey(key)!));
+    for (const key of expr.keys) getFreeVars(expr.getKey(key)!, set);
 
-  return result;
+  return;
 }
 
-export function getSymbols(
-  expr: BoxedExpression,
-  set: Set<string>
-): Set<string> {
+export function getSymbols(expr: BoxedExpression, set: Set<string>): void {
   if (expr.symbol) {
     set.add(expr.symbol);
-    return set;
+    return;
   }
 
-  if (!expr.ops && !expr.keys) return set;
+  if (!expr.ops && !expr.keys) return;
 
   if (expr.ops) for (const op of expr.ops) getSymbols(op, set);
 
   if (expr.keys)
     for (const key of expr.keys) getSymbols(expr.getKey(key)!, set);
 
-  return set;
+  return;
 }
 
 export function getSubexpressions(